Move collections editor out of varexp and into widgets
