+++ spyder/utils/environ.py
@@ -16,8 +16,7 @@ from qtpy.QtWidgets import QDialog, QMessageBox
 
 # Local imports
 from spyder.config.base import _
-from spyder.plugins.variableexplorer.widgets.collectionseditor import (
-        CollectionsEditor)
+from spyder.widgets.collectionseditor import CollectionsEditor
 from spyder.py3compat import PY2, iteritems, to_text_string, to_binary_string
 from spyder.utils import icon_manager as ima
 from spyder.utils.encoding import to_unicode_from_fs
