+++ spyder/widgets/collectionseditor.py
